+++ objectbox/CHANGELOG.md
@@ -8,7 +8,9 @@
   * renamed the former `putAsync` to `putQueuedAwaitResult`,
   * new `putAndGetAsync` and `putAndGetManyAsync` which return a copy of the given objects with new
     IDs set.
-* Add new async `Query` APIs: `findAsync`, `findFirstAsync`, `findUniqueAsync` and `findIdsAsync`.
+  * new `removeAsync`, `removeManyAsync` and `removeAllAsync`.
+* Add new async `Query` APIs: `findAsync`, `findFirstAsync`, `findUniqueAsync`, `findIdsAsync` and 
+  `removeAsync`.
 * Support sending objects containing `ToOne` and `ToMany` across isolates, e.g. when using
   `store.runInTransactionAsync`. #340
 * `Store.attach` (and `Store.fromReference`) do not longer accept a null model, which was not